requirements.txt
flask
waitress
reportlab
werkzeug
pywebview
